Crate boot_info_x86_64

Modules

kernel_map
This module contains constants that define how the kernel address space is laid out on x86_64. The 511th P4 entry (virtual addresses 0xffff_ff80_0000_0000 through 0xffff_ffff_ffff_ffff) is always mapped to the kernel P3. The rest of the virtual address space (virtual addresses 0x0000_0000_0000_0000 through 0xffff_ff7f_ffff_ffff are free for userspace to use).

Structs

BootInfo

LoadedImage
Describes an image loaded from the filesystem by the loader, as the kernel does not have the capabilities to do so. Images are expected to have three segments (rodata loaded as read-only, data loaded as read+write, and text loaded as read+execute).

LoadedImages

MemoryMap

MemoryMapEntry

Segment

VideoModeInfo

Enums

MemoryType

PixelFormat

Constants

BOOT_INFO_MAGIC

MAX_CAPABILITY_STREAM_LENGTH

MAX_IMAGE_LOADED_SEGMENTS

MAX_IMAGE_NAME_LENGTH
This is one less than a power-of-two, because then it's aligned when placed after the length byte.

MAX_LOADED_IMAGES

MAX_MEMORY_MAP_ENTRIES