Enum boot_info_x86_64::MemoryType

#[repr(C)]
pub enum MemoryType {
    Conventional,
    AcpiReclaimable,
    LoadedImage,
    KernelPageTables,
    KernelHeap,
    BootInfo,
}

Variants

Conventional

Memory that can be used freely by the OS.

AcpiReclaimable

Memory that contains ACPI tables. After the OS has finished with them, this may be treated as conventional memory.

LoadedImage

Memory occupied by images that the loader has been asked to load from disk. If the kernel can determine that an image is no longer needed, it may use this memory.

KernelPageTables

Memory that is occupied by page tables created by the loader for the kernel. If the kernel can determine that it no longer needs part of this mapped, it may use this memory.

KernelHeap

Memory that has been mapped for the kernel heap.

BootInfo

Memory that is occupied by the boot info constructed by the loader for the kernel. Contains the BootInfo structure, and all the structures that are referenced by it. After the kernel has finished with this data, it may use this memory.